add menu tv show

+++ Moviers/app/src/main/java/com/rivaldofez/moviers/entity/MovieEntity.kt
@@ -4,12 +4,12 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class MovieEntity (
+data class MovieEntity(
     var id: Int,
     var title: String,
     var overview: String,
     var date: String,
-    var rating: Int,
+    var rating: Float,
     var studio: String,
     var posterPath: String
 ):Parcelable
